+++ Common/Enum/FileExtension.cs
@@ -1,0 +1,12 @@
+using System.Runtime.Serialization;
+namespace Common.Enum;
+
+public enum FileExtension
+{
+    LOG,
+    PIRATE,
+    JSON,
+    BIN,
+    gitignore,
+    gitattributes
+}