+++ Common/FileHandlers/FileReadHandler.cs
@@ -1,18 +1,19 @@
 using System.Globalization;
+using Common.Enum;
 using Common.FileHandlers.Interfaces;
 
 namespace Common.FileHandlers;
 
 [Serializable]
-public class FileReadHandler : IFileReadHandler
+public class FileReadHandler : BaseFileHandler, IFileReadHandler
 {
-    public async Task<string> ReadAllTextFromFile(string name, string extension, string location)
+    public async Task<string> ReadAllTextFromFile(string name, FileExtension extension, string location)
     {
-        if (name == string.Empty || extension == string.Empty) { throw new ArgumentNullException("Name, Text or Extension provided is empty"); }
+        if (name == string.Empty) { throw new ArgumentNullException("Name, Text or Extension provided is empty"); }
 
-        name += extension;
+        var nameAndExtension = name + GetFileExtension(extension);
         var targetFolder = Path.Combine(Environment.CurrentDirectory, location);
-        string fileName = Path.Combine(targetFolder, name);
+        string fileName = Path.Combine(targetFolder, nameAndExtension);
 
         if (FileExists(name, extension, location))
         {
@@ -24,11 +25,11 @@ public class FileReadHandler : IFileReadHandler
         }
     }
 
-    public bool FileExists(string name, string extension, string location)
+    public bool FileExists(string name, FileExtension extension, string location)
     {
-        if (name == string.Empty || extension == string.Empty) { throw new ArgumentNullException("Name, Text or Extension provided is empty"); }
+        if (name == string.Empty) { throw new ArgumentNullException("Name, Text or Extension provided is empty"); }
 
-        name += extension;
+        name += GetFileExtension(extension);
         var targetFolder = Path.Combine(Environment.CurrentDirectory, location);
         string fileName = Path.Combine(targetFolder, name);