+++ PirateInterpreter.Test/InterpretersTest.cs
@@ -1,5 +1,8 @@
+using System.Collections.Generic;
 using PirateInterpreter.Interpreters;
+using PirateInterpreter.Values;
 using PirateParser.Node;
+using PirateParser.Node.Interfaces;
 
 namespace PirateInterpreter.Test
 {
@@ -20,8 +23,9 @@ namespace PirateInterpreter.Test
             var result = interpreter.VisitNode();
 
             // Assert
-            Assert.IsType<Values.Integer>(result);
-            Assert.Equal(2, result.Value);
+            Assert.IsType<List<BaseValue>>(result);
+            Assert.IsType<Values.Integer>(result[0]);
+            Assert.Equal(2, result[0].Value);
         }
 
         [Fact]
@@ -39,8 +43,9 @@ namespace PirateInterpreter.Test
             var result = interpreter.VisitNode();
 
             // Assert
-            Assert.IsType<Values.Boolean>(result);
-            Assert.Equal(1, result.Value);
+            Assert.IsType<List<BaseValue>>(result);
+            Assert.IsType<Values.Boolean>(result[0]);
+            Assert.Equal(1, result[0].Value);
         }
 
         [Fact]
@@ -54,8 +59,9 @@ namespace PirateInterpreter.Test
             var result = interpreter.VisitNode();
 
             // Assert
-            Assert.IsType<Values.Integer>(result);
-            Assert.Equal(1, result.Value);
+            Assert.IsType<List<BaseValue>>(result);
+            Assert.IsType<Values.Integer>(result[0]);
+            Assert.Equal(1, result[0].Value);
         }
 
         [Fact]
@@ -73,8 +79,9 @@ namespace PirateInterpreter.Test
             var result = interpreter.VisitNode();
 
             // Assert
-            Assert.IsType<Values.Variable>(result);
-            Assert.Equal(1, result.Value);
+            Assert.IsType<List<BaseValue>>(result);
+            Assert.IsType<Values.Variable>(result[0]);
+            Assert.Equal(1, result[0].Value);
         }
 
         [Fact]
@@ -96,8 +103,9 @@ namespace PirateInterpreter.Test
             var result = interpreter.VisitNode();
 
             // Assert
-            Assert.IsType<Values.Variable>(result);
-            Assert.Equal(2, result.Value);
+            Assert.IsType<List<BaseValue>>(result);
+            Assert.IsType<Values.Variable>(result[0]);
+            Assert.Equal(2, result[0].Value);
         }
 
         [Fact]
@@ -119,8 +127,132 @@ namespace PirateInterpreter.Test
             var result = interpreter.VisitNode();
 
             // Assert
-            Assert.IsType<Values.Variable>(result);
-            Assert.Equal(1, result.Value);
+            Assert.IsType<List<BaseValue>>(result);
+            Assert.IsType<Values.Variable>(result[0]);
+            Assert.Equal(1, result[0].Value);
+        }
+
+        [Fact]
+        public void ShouldInterpretIfStatementNode()
+        {
+            // Arrange
+            var ifStatementNode = new IfStatementNode(
+                new ComparisonOperationNode(
+                    new ValueNode(new Token(TokenGroup.VALUE, TokenValue.INT, 1)),
+                    new Token(TokenGroup.COMPARISONOPERATORS, TokenComparisonOperators.DOUBLEEQUALS),
+                    new ValueNode(new Token(TokenGroup.VALUE, TokenValue.INT, 1))
+                ),
+                new List<INode>() {
+                    new VariableAssignNode(
+                        new Token(TokenGroup.TYPEKEYWORD, TokenTypeKeyword.VAR),
+                        new ValueNode(new Token(TokenGroup.SYNTAX, TokenSyntax.IDENTIFIER, "a")),
+                        new ValueNode(new Token(TokenGroup.VALUE, TokenValue.INT, 1))
+                    )
+                }
+            );
+
+            // Act
+            var interpreter = new IfStatementNodeInterpreter(ifStatementNode, new InterpreterFactory(), A.Fake<ILogger>());
+            var result = interpreter.VisitNode();
+
+            // Assert
+            Assert.IsType<List<BaseValue>>(result);
+            Assert.IsType<Values.Variable>(result[0]);
+            Assert.Equal(1, result[0].Value);
+        }
+
+        [Fact]
+        public void ShouldInterpretIfStatementNodeWithElse()
+        {
+            // Arrange
+            var ifStatementNode = new IfStatementNode(
+                new ComparisonOperationNode(
+                    new ValueNode(new Token(TokenGroup.VALUE, TokenValue.INT, 1)),
+                    new Token(TokenGroup.COMPARISONOPERATORS, TokenComparisonOperators.DOUBLEEQUALS),
+                    new ValueNode(new Token(TokenGroup.VALUE, TokenValue.INT, 2))
+                ),
+                new List<INode>() {
+                    new VariableAssignNode(
+                        new Token(TokenGroup.TYPEKEYWORD, TokenTypeKeyword.VAR),
+                        new ValueNode(new Token(TokenGroup.SYNTAX, TokenSyntax.IDENTIFIER, "a")),
+                        new ValueNode(new Token(TokenGroup.VALUE, TokenValue.INT, 1))
+                    )
+                },
+                new List<INode>() {
+                    new VariableAssignNode(
+                        new Token(TokenGroup.TYPEKEYWORD, TokenTypeKeyword.VAR),
+                        new ValueNode(new Token(TokenGroup.SYNTAX, TokenSyntax.IDENTIFIER, "a")),
+                        new ValueNode(new Token(TokenGroup.VALUE, TokenValue.INT, 2))
+                    )
+                }
+            );
+
+            // Act
+            var interpreter = new IfStatementNodeInterpreter(ifStatementNode, new InterpreterFactory(), A.Fake<ILogger>());
+            var result = interpreter.VisitNode();
+
+            // Assert
+            Assert.IsType<List<BaseValue>>(result);
+            Assert.IsType<Values.Variable>(result[0]);
+            Assert.Equal(2, result[0].Value);
+        }
+
+        // [Fact]
+        // public void ShouldInterpretWhileLoopStatementNode()
+        // {
+        //     // Arrange
+        //     var whileStatementNode = new WhileLoopStatementNode(
+        //         new ComparisonOperationNode(
+        //             new ValueNode(new Token(TokenGroup.VALUE, TokenValue.INT, 1)),
+        //             new Token(TokenGroup.COMPARISONOPERATORS, TokenComparisonOperators.DOUBLEEQUALS),
+        //             new ValueNode(new Token(TokenGroup.VALUE, TokenValue.INT, 1))
+        //         ),
+        //         new List<INode>() {
+        //             new VariableAssignNode(
+        //                 new Token(TokenGroup.TYPEKEYWORD, TokenTypeKeyword.VAR),
+        //                 new ValueNode(new Token(TokenGroup.SYNTAX, TokenSyntax.IDENTIFIER, "a")),
+        //                 new ValueNode(new Token(TokenGroup.VALUE, TokenValue.INT, 1))
+        //             )
+        //         }
+        //     );
+
+        //     // Act
+        //     var interpreter = new WhileLoopStatementNodeInterpreter(whileStatementNode, new InterpreterFactory(), A.Fake<ILogger>());
+        //     var result = interpreter.VisitNode();
+
+        //     // Assert
+        //     Assert.IsType<Values.Variable>(result);
+        //     Assert.Equal(1, result[0].Value);
+        // }
+
+        [Fact]
+        public void ShouldInterpretForLoopStatementNode()
+        {
+            // Arrange
+            var forStatementNode = new ForLoopStatementNode(
+                new VariableAssignNode(
+                    new Token(TokenGroup.TYPEKEYWORD, TokenTypeKeyword.VAR),
+                    new ValueNode(new Token(TokenGroup.SYNTAX, TokenSyntax.IDENTIFIER, "a")),
+                    new ValueNode(new Token(TokenGroup.VALUE, TokenValue.INT, 0))
+                ),
+                new ValueNode(new Token(TokenGroup.VALUE, TokenValue.INT, 10)),
+                new List<INode>() {
+                    new VariableAssignNode(
+                        new Token(TokenGroup.TYPEKEYWORD, TokenTypeKeyword.VAR),
+                        new ValueNode(new Token(TokenGroup.SYNTAX, TokenSyntax.IDENTIFIER, "a")),
+                        new ValueNode(new Token(TokenGroup.VALUE, TokenValue.INT, 1))
+                    )
+                }
+            );
+
+            // Act
+            var interpreter = new ForLoopStatementNodeInterpreter(forStatementNode, new InterpreterFactory(), A.Fake<ILogger>());
+            var result = interpreter.VisitNode();
+
+            // Assert
+            Assert.IsType<List<BaseValue>>(result);
+            Assert.IsType<Values.Variable>(result[0]);
+            Assert.Equal(1, result[0].Value);
         }
     }
 }