+++ PirateParser/Parsers/ParserFactory.cs
@@ -4,10 +4,16 @@ namespace PirateParser.Parsers;
 
 public class ParserFactory: IParserFactory
 {
-    public ITokenParser GetParser(Token token, List<Token> tokens, ILogger logger)
+    public BaseParser GetParser(Token token, List<Token> tokens, ILogger logger)
     {
         switch(token.TokenType)
         {
+            case TokenControlKeyword.IF:
+                return new IfStatementParser(tokens, token, logger, this);
+            case TokenControlKeyword.WHILE:
+                return new WhileLoopStatementParser(tokens, token, logger, this);
+            case TokenControlKeyword.FOR:
+                return new ForLoopStatementParser(tokens, token, logger, this);
             case TokenTypeKeyword:
                 return new VariableAssignParser(tokens, token, logger, this);